requirements.txt
fastapi
uvicorn
langchain
langchain-community 
python-dotenv
langchain-huggingface
faiss-cpu
jinja2
wikipedia
pypdf
sentence-transformers 
torch                 
transformers          
accelerate